clean up analyzer logging

+++ telemetry/analyzer/src/main/java/ru/yandex/practicum/telemetry/analyzer/handler/ScenarioRemovedHandler.java
@@ -1,7 +1,6 @@
 package ru.yandex.practicum.telemetry.analyzer.handler;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.kafka.telemetry.event.HubEventAvro;
 import ru.yandex.practicum.kafka.telemetry.event.ScenarioRemovedEventAvro;
@@ -9,7 +8,6 @@ import ru.yandex.practicum.telemetry.analyzer.repository.ActionRepository;
 import ru.yandex.practicum.telemetry.analyzer.repository.ConditionRepository;
 import ru.yandex.practicum.telemetry.analyzer.repository.ScenarioRepository;
 
-@Slf4j
 @Component
 @RequiredArgsConstructor
 public class ScenarioRemovedHandler implements HubEventHandler {
@@ -20,12 +18,15 @@ public class ScenarioRemovedHandler implements HubEventHandler {
     @Override
     public void handle(HubEventAvro event) {
         ScenarioRemovedEventAvro payload = (ScenarioRemovedEventAvro) event.getPayload();
+        if (payload == null) {
+            return;
+        }
+
         scenarioRepository.findByHubIdAndName(event.getHubId(), payload.getName())
                 .ifPresent(scenario -> {
                     actionRepository.deleteByScenario(scenario);
                     conditionRepository.deleteByScenario(scenario);
                     scenarioRepository.delete(scenario);
-                    log.info("Removed scenario {} for hub {}", payload.getName(), event.getHubId());
                 });
     }
